requirements.txt
fastapi
uvicorn
line-bot-sdk
openai
python-dotenv
